add test-path to AccountStatisticsServiceGrpcHost script

+++ resources/windows/Deploy/AccountStatisticsService/AccountStatisticsServiceGrpcHost.ps1
@@ -10,13 +10,18 @@ $assGrpcPort = 5038
 
 Write-host "[INFO] Start ${serviceName} deploy script"
 Write-Host -ForegroundColor Green "[INFO] Edit $pathtojson"
-$jsonAppsetings = Get-Content -Raw -path $pathtojson  | % { $_ -replace '[\s^]//.*', "" } | ConvertFrom-Json 
+if (test-path $pathtojson){
+    $jsonAppsetings = Get-Content -Raw -path $pathtojson  | % { $_ -replace '[\s^]//.*', "" } | ConvertFrom-Json 
 
 # Настраиваем секцию подключения к БД
-$jsonAppsetings.ConnectionStrings.AccountStatisticsServiceDb = "data source=localhost;initial catalog=AccountStatisticsService;Integrated Security=true;MultipleActiveResultSets=True;"
+    $jsonAppsetings.ConnectionStrings.AccountStatisticsServiceDb = "data source=localhost;initial catalog=AccountStatisticsService;Integrated Security=true;MultipleActiveResultSets=True;"
 
 # Меняем порт GrpcPort
-$jsonAppsetings.Kestrel.EndPoints.Https.Url = "http://localhost:${assGrpcPort}"
+    $jsonAppsetings.Kestrel.EndPoints.Https.Url = "http://localhost:${assGrpcPort}"
 
-ConvertTo-Json $jsonAppsetings -Depth $jsonDepth  | Format-Json | Set-Content $pathtojson -Encoding UTF8
-Write-Host -ForegroundColor Green "[INFO] $pathtojson renewed with json depth $jsonDepth"
+    ConvertTo-Json $jsonAppsetings -Depth $jsonDepth  | Format-Json | Set-Content $pathtojson -Encoding UTF8
+    Write-Host -ForegroundColor Green "[INFO] $pathtojson renewed with json depth $jsonDepth"
+}
+else{
+    Write-Host -ForegroundColor Green "[INFO] $pathtojson not exists"
+}
